Re enable llvm-src.sh

+++ deps/install_scripts/all.sh
@@ -5,10 +5,10 @@
 # #SBATCH -o %x-%j.txt
 # #SBATCH -e %x-%j.txt
 
-# if [[ $HOSTNAME == login* ]]; then
-#     echo "DO NOT RUN THIS ON LOGIN NODE. Use `sbatch .../all.sh`"
-#     exit 1
-# fi
+if [[ $HOSTNAME == login* ]]; then
+    echo "DO NOT RUN THIS ON LOGIN NODE. Use `sbatch .../all.sh`"
+    exit 1
+fi
 
 TADASHI_ROOT="$(git rev-parse --show-toplevel)"
 rm -rf "${TADASHI_DEPS_PREFIX:-$TADASHI_ROOT/deps/opt}"
@@ -18,7 +18,7 @@ source "$TADASHI_ROOT/scripts/genoa.source"
 
 SCRIPTS_DIR=${TADASHI_DEPS_PREFIX:-$TADASHI_ROOT/deps/install_scripts}
 "${SCRIPTS_DIR}/gmp.sh"
-"${SCRIPTS_DIR}/llvm.sh"
+"${SCRIPTS_DIR}/llvm_src.sh"
 "${SCRIPTS_DIR}/python.sh"
 "${SCRIPTS_DIR}/swig.sh"
